refactor: testUtils 파일을 renderWithProviders로 변경하고 위치 이동
