+++ frontend/.gitignore
@@ -5,3 +5,4 @@ build-storybook.log
 .env
 testOutDir
 junit.xml
+coverage
